fix particle number overflow for large in_prt_n2max

+++ sourcemod.f
@@ -3,7 +3,7 @@ c     ----------------
       implicit none
       integer :: src_ns,src_ninit
       integer :: src_nsurf,src_nnonth,src_nnew,src_ninitnew
-      integer :: src_nvacantmin=huge(src_ns)
+      integer*8 :: src_nvacantmin=huge(src_ns)
       integer :: src_nflux=0
 c
       integer,allocatable :: src_ivacant(:) !array of vacant particle array locations
